created permissions manager api

+++ src/main/java/com/scolastico/discord_exe/event/events/CommandW2G.java
@@ -3,11 +3,13 @@ package com.scolastico.discord_exe.event.events;
 import com.scolastico.discord_exe.Disc0rd;
 import com.scolastico.discord_exe.etc.ErrorHandler;
 import com.scolastico.discord_exe.etc.Tools;
+import com.scolastico.discord_exe.etc.permissions.PermissionsManager;
 import com.scolastico.discord_exe.event.EventRegister;
 import com.scolastico.discord_exe.event.handlers.CommandHandler;
 import com.scolastico.discord_exe.event.handlers.EventHandler;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.JDA;
+import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
@@ -21,45 +23,54 @@ public class CommandW2G implements EventHandler,CommandHandler {
     @Override
     public boolean respondToCommand(String cmd, String[] args, JDA jda, MessageReceivedEvent event, long senderId, long serverId) {
         if (cmd.equalsIgnoreCase("w2g") || cmd.equalsIgnoreCase("watch2gether")) {
-            try {
-                if (args.length == 0 || args.length == 1) {
-                    event.getMessage().delete().queue();
-                    EmbedBuilder embedBuilder = new EmbedBuilder();
-                    String jsonResponse;
-                    if (args.length == 0) {
-                        jsonResponse = Tools.getInstance().sendPostRequest(
-                                "https://www.watch2gether.com/rooms/create.json",
-                                "{\"share\":\"" + Disc0rd.getConfig().getW2gDefaultPlayback().replaceAll("(\\\\|\")", "") + "\",\"api_key\":\"" + Disc0rd.getConfig().getW2gToken() + "\",\"bg_color\":\"#23272A\",\"bg_opacity\":\"100\"}"
-                        );
-                    } else {
-                        jsonResponse = Tools.getInstance().sendPostRequest(
-                                "https://www.watch2gether.com/rooms/create.json",
-                                "{\"share\":\"" + args[0].replaceAll("(\\\\|\")", "") + "\",\"api_key\":\"" + Disc0rd.getConfig().getW2gToken() + "\",\"bg_color\":\"#23272A\",\"bg_opacity\":\"100\"}"
-                        );
-                    }
-                    if (jsonResponse != null) {
-                        Object obj = new JSONParser().parse(jsonResponse);
-                        if (obj instanceof JSONObject) {
-                            Object streamKeyObj = ((JSONObject) obj).get("streamkey");
-                            if (streamKeyObj instanceof String) {
-                                String streamKey = (String) streamKeyObj;
-                                embedBuilder.setColor(Color.green);
-                                embedBuilder.setTitle("Watch2Gether");
-                                embedBuilder.setDescription("Your Watch2Gether room was created!\nYou can access it with the following url:\n\n<https://watch2gether.com/rooms/" + streamKey + ">");
-                                embedBuilder.setThumbnail("https://www.watch2gether.com/static/w2g-oglogo.png");
-                                event.getChannel().sendMessage(embedBuilder.build()).queue();
-                                return true;
+            EmbedBuilder embedBuilder = new EmbedBuilder();
+            Member member = event.getGuild().getMember(event.getAuthor());
+            if (member == null) return false;
+            if (PermissionsManager.getInstance().checkPermission(event.getGuild(), member, "w2g")) {
+                try {
+                    if (args.length == 0 || args.length == 1) {
+                        event.getMessage().delete().queue();
+                        String jsonResponse;
+                        if (args.length == 0) {
+                            jsonResponse = Tools.getInstance().sendPostRequest(
+                                    "https://www.watch2gether.com/rooms/create.json",
+                                    "{\"share\":\"" + Disc0rd.getConfig().getW2gDefaultPlayback().replaceAll("(\\\\|\")", "") + "\",\"api_key\":\"" + Disc0rd.getConfig().getW2gToken() + "\",\"bg_color\":\"#23272A\",\"bg_opacity\":\"100\"}"
+                            );
+                        } else {
+                            jsonResponse = Tools.getInstance().sendPostRequest(
+                                    "https://www.watch2gether.com/rooms/create.json",
+                                    "{\"share\":\"" + args[0].replaceAll("(\\\\|\")", "") + "\",\"api_key\":\"" + Disc0rd.getConfig().getW2gToken() + "\",\"bg_color\":\"#23272A\",\"bg_opacity\":\"100\"}"
+                            );
+                        }
+                        if (jsonResponse != null) {
+                            Object obj = new JSONParser().parse(jsonResponse);
+                            if (obj instanceof JSONObject) {
+                                Object streamKeyObj = ((JSONObject) obj).get("streamkey");
+                                if (streamKeyObj instanceof String) {
+                                    String streamKey = (String) streamKeyObj;
+                                    embedBuilder.setColor(Color.green);
+                                    embedBuilder.setTitle("Watch2Gether");
+                                    embedBuilder.setDescription("Your Watch2Gether room was created!\nYou can access it with the following url:\n\n<https://watch2gether.com/rooms/" + streamKey + ">");
+                                    embedBuilder.setThumbnail("https://www.watch2gether.com/static/w2g-oglogo.png");
+                                    event.getChannel().sendMessage(embedBuilder.build()).queue();
+                                    return true;
+                                }
                             }
                         }
+                        embedBuilder.setColor(Color.red);
+                        embedBuilder.setTitle("Sorry,");
+                        embedBuilder.setDescription("but an unknown error occurred! Please try it again in a few minutes!");
+                        event.getChannel().sendMessage(embedBuilder.build()).queue();
+                        return true;
                     }
-                    embedBuilder.setColor(Color.red);
-                    embedBuilder.setTitle("Sorry,");
-                    embedBuilder.setDescription("but an unknown error occurred! Please try it again in a few minutes!");
-                    event.getChannel().sendMessage(embedBuilder.build()).queue();
-                    return true;
+                } catch (Exception e) {
+                    ErrorHandler.getInstance().handle(e);
                 }
-            } catch (Exception e) {
-                ErrorHandler.getInstance().handle(e);
+            } else {
+                embedBuilder.setColor(Color.RED);
+                embedBuilder.setTitle("Sorry,");
+                embedBuilder.setDescription("but you dont have the permission to use this command!");
+                event.getChannel().sendMessage(embedBuilder.build()).queue();
             }
         }
         return false;
@@ -88,7 +99,7 @@ public class CommandW2G implements EventHandler,CommandHandler {
 
     @Override
     public void registerEvents(EventRegister eventRegister) {
-        CommandW2G element = new CommandW2G();
-        eventRegister.registerCommand(element);
+        eventRegister.registerCommand(this);
+        PermissionsManager.getInstance().registerPermission("w2g", "Allow a user to use the watch 2 gether command.", true);
     }
 }
